% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdbcluster_package.R
\name{run}
\alias{run}
\title{Main function}
\usage{
run(file_name = "", file_type = "list", save_files = F, skip_frames = 0,
  plots = T, protein_representation = "CA", chain = "",
  cluster_quantity = 10, receptor_peptide_mode = F,
  reference_structure_pdb_code = "", reference_structure_file = "",
  reference_chain_id = "", kmeans_algorithm = "Hartigan-Wong",
  output_directory = "")
}
\arguments{
\item{file_name}{path to file}

\item{file_type}{list or file}

\item{save_files}{boolean}

\item{skip_frames}{number how many first frames should be skipped}

\item{plots}{boolean}

\item{protein_representation}{CA or all}

\item{chain}{chain ID}

\item{cluster_quantity}{number}

\item{receptor_peptide_mode}{boolean}

\item{reference_structure_pdb_code}{code from PDB web repository}

\item{reference_structure_file}{path to file}

\item{reference_chain_id}{chain ID}

\item{kmeans_algorithm}{Hartigan-Wong / MacQueen / Lloyd / Forgy}

\item{output_directory}{path to directory}
}
\value{
return object with clusters data
}
\description{
Main function
}
